+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGKXJK585XDG40BTHQNK37P3"
+        "01KGM98JAFA4K8J9SVBDJYK8YR"
     )
     r"""The unique identifier of the data source"""